Draggable Windows in the BWM
We're one step closer to the full Basic Window Manager implementation.
I've created a UserControl called "TitleBar" for the BWM. This
UserControl contains the window handling code for draggable windows, as
well as a design-time settable AppName property that allows you to
change the application display name (what text is shown on the
titlebar).

Shiftorium upgrades to unlock Draggable Windows will be added soon, but
I have to clean the beast.