Shiftorium - Draggable Windows upgrade + Bugfixes
Alright, I've added the Draggable Windows upgrade to the Shiftorium, as
well as fixed a quick bug causing "No Items Available" to display when
there were items clearly available.

I've also fixed a positioning bug with BWM windows causing them to start
in the center of the screen, relative to the titlebar, rather than the
form.

I've also re-arranged the code a bit, organizing files in folders. APIs
are in the APIs folder, Desktop Environments are in the Desktop
Environments folder, etc.

The shiftorium still doesn't have draggable window functionality yet...