BWM - Skinning!
Yay! ShiftOS Next can finally be Shifted! Yes, that's true. I've added
basic skinning to the Basic Window Manager. Currently, the only way to
get to it is by typing "open shifter" in the Terminal, and the Shifter
will open. Choose the settings you want and hit Apply!

This is also the Alpha 3 commit, so after this I'll make the next
release.